+++ template/CentOS-8/scripts/cleanup.sh
@@ -6,14 +6,15 @@ umask 0022
 ## Network configuration
 ## ======================================================================
 
+systemctl start NetworkManager.service
 nmcli radio all off
 
 systemctl stop NetworkManager.service
 rm -r /var/lib/NetworkManager/*
 
-mkdir /etc/sysconfig/network-scripts/dist
+mkdir -p /etc/sysconfig/network-scripts/dist
 for ifcfg in /etc/sysconfig/network-scripts/ifcfg-*; do
-  [[ ${ifcfg##*/} == ifcfg-lo ]] && continue
+  [[ ! -f $ifcfg || ${ifcfg##*/} == ifcfg-lo ]] && continue
   mv "$ifcfg" /etc/sysconfig/network-scripts/dist/
 done
